+++ addons/lightdash/app/static/scripts.js
@@ -1,0 +1,28 @@
+// ### Necessary element styling 
+document.querySelectorAll(".icey_text_fit").forEach(el => {
+	el.style.display = "inline-block";
+	el.style.whiteSpace = "nowrap";
+});
+
+// ### Set individual font sizes
+let icey_textFit_timeout = null;
+function icey_textFit() {
+	clearTimeout(icey_textFit_timeout);
+	icey_textFit_timeout = setTimeout(function () {
+		document.querySelectorAll(".icey_text_fit").forEach(el => {
+			if (el.innerText.length === 0) return;
+
+			const parentWidth = el.parentElement.offsetWidth;
+			
+			let fontSize = parentWidth / el.innerText.length;
+			el.style.setProperty("--font_size", fontSize);
+
+			const currentWidth = el.offsetWidth;
+			if (currentWidth > 0) {
+				const ratio = currentWidth / parentWidth;
+				fontSize = fontSize / ratio;
+				el.style.setProperty("--font_size", fontSize);
+			}
+		});
+	}, 10);
+}